+++ src/bika/lims/browser/worksheet/views/add_analyses.py
@@ -66,7 +66,6 @@ class AddAnalysesView(BikaListingView):
         self.contentFilter = {
             "portal_type": "Analysis",
             "review_state": "unassigned",
-            "isSampleReceived": True,
             "sort_on": "getPrioritySortkey",
         }